refactor: rename redux-states/*/action-types.ts to model.ts
