+++ src/redux-states/firebase/model.ts
