+++ src/redux-states/project/reducer.ts
@@ -13,7 +13,7 @@ import {
   UPDATE_PROJECT,
   WRITE_PROJECT_FAILURE,
 } from './action-constants';
-import { IProjectState } from './action-types';
+import { IProjectState } from './model';
 
 const initialState: IProjectState = {
   projects: {} as IProjects<IProject>,
