+++ src/redux-states/task-list/index.ts
@@ -1,6 +1,6 @@
 import * as taskListActions from './actions';
 import * as taskListConstants from './action-constants';
-import * as taskListTypes from './action-types';
+import * as taskListTypes from './model';
 import taskListReducer from './reducer';
 
 export { taskListActions, taskListConstants, taskListReducer, taskListTypes };
