+++ src/redux-states/user/actions.ts
@@ -2,7 +2,7 @@ import * as AppTypes from 'Entities/types';
 import * as StateTypes from 'States/types';
 import IUser from 'Entities/user';
 import { auth } from 'Utils/firebase';
-import { IUserState } from './action-types';
+import { IUserState } from './model';
 import { alertSuccess, alertError } from './alert-action-creators';
 import {
   REGISTER_REQUEST,
