+++ src/redux-states/user/model.ts
