+++ src/redux-states/user/reducer.ts
@@ -1,7 +1,7 @@
 import * as AppTypes from 'Entities/types';
 import IUser from 'Entities/user';
 import * as StateTypes from 'States/types';
-import { IUserState } from 'States/user/action-types';
+import { IUserState } from 'States/user/model';
 import { ALERT_SUCCESS, ALERT_ERROR, ALERT_CLEAR } from './alert-action-constants';
 import {
   REGISTER_SUCCESS,
